Fixed Windows build, fixed some compiler warnings Fixed crash on subresources data loading in mixed order.

+++ Modules/Data/Types/Include/Methane/Data/Parallel.hpp
@@ -34,14 +34,14 @@ void ParallelFor(const Iterator& begin_it, const Iterator& end_it, UnaryFunction
 {
     ITT_FUNCTION_TASK();
 
-    const size_t   items_count = std::distance(begin_it, end_it);
+    const uint32_t items_count = static_cast<uint32_t>(std::distance(begin_it, end_it));
     const uint32_t thread_count = std::thread::hardware_concurrency();
     const uint32_t group_size = std::max(1u, static_cast<uint32_t>(std::ceil(static_cast<float>(items_count) / thread_count)));
 
     std::vector<std::thread> threads;
     threads.reserve(thread_count);
 
-    for (Iterator it = begin_it; std::distance(begin_it, it) < items_count; it += group_size)
+    for (Iterator it = begin_it; static_cast<uint32_t>(std::distance(begin_it, it)) < items_count; it += group_size)
     {
         threads.push_back(
             std::thread([=, &f]()
